requirements.txt
numpy
scipy
pandas
pyyaml
pydantic